demo/requirements.txt
# Minimal requirements for Hackathon Demo
streamlit>=1.28.0
openai>=2.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
